small

+++ qlat-utils/qlat_utils/include/qlat-utils/types.h
@@ -711,7 +711,7 @@ qacc bool get_data_type_is_double()
   DATA_TYPE cur = get_data_type<M>();
   if ((cur < FLOATIND) or (cur == INVALID_TYPE)) {
     if (get_id_node() == 0) {
-      printf("Given type not float/double %d \n", cur);
+      printf("Given type not float/double %d \n", (int)cur);
     }
     qassert(false);
   }
